MGS-7377 [#mali] VX504X08X-SW-99002-r45p0-01eac0

+++ drivers/gpu/arm/midgard/mali_kbase_mem_pool_debugfs.c
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0 WITH Linux-syscall-note
 /*
  *
- * (C) COPYRIGHT 2014-2022 ARM Limited. All rights reserved.
+ * (C) COPYRIGHT 2014-2023 ARM Limited. All rights reserved.
  *
  * This program is free software and is provided to you under the terms of the
  * GNU General Public License version 2 as published by the Free Software
@@ -25,25 +25,21 @@
 #include "mali_kbase_mem_pool_debugfs.h"
 #include "mali_kbase_debugfs_helper.h"
 
-void kbase_mem_pool_debugfs_trim(void *const array, size_t const index,
-	size_t const value)
+void kbase_mem_pool_debugfs_trim(void *const array, size_t const index, size_t const value)
 {
 	struct kbase_mem_pool *const mem_pools = array;
 
-	if (WARN_ON(!mem_pools) ||
-		WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
+	if (WARN_ON(!mem_pools) || WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
 		return;
 
 	kbase_mem_pool_trim(&mem_pools[index], value);
 }
 
-void kbase_mem_pool_debugfs_set_max_size(void *const array,
-	size_t const index, size_t const value)
+void kbase_mem_pool_debugfs_set_max_size(void *const array, size_t const index, size_t const value)
 {
 	struct kbase_mem_pool *const mem_pools = array;
 
-	if (WARN_ON(!mem_pools) ||
-		WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
+	if (WARN_ON(!mem_pools) || WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
 		return;
 
 	kbase_mem_pool_set_max_size(&mem_pools[index], value);
@@ -53,8 +49,7 @@ size_t kbase_mem_pool_debugfs_size(void *const array, size_t const index)
 {
 	struct kbase_mem_pool *const mem_pools = array;
 
-	if (WARN_ON(!mem_pools) ||
-		WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
+	if (WARN_ON(!mem_pools) || WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
 		return 0;
 
 	return kbase_mem_pool_size(&mem_pools[index]);
@@ -64,32 +59,28 @@ size_t kbase_mem_pool_debugfs_max_size(void *const array, size_t const index)
 {
 	struct kbase_mem_pool *const mem_pools = array;
 
-	if (WARN_ON(!mem_pools) ||
-		WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
+	if (WARN_ON(!mem_pools) || WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
 		return 0;
 
 	return kbase_mem_pool_max_size(&mem_pools[index]);
 }
 
-void kbase_mem_pool_config_debugfs_set_max_size(void *const array,
-	size_t const index, size_t const value)
+void kbase_mem_pool_config_debugfs_set_max_size(void *const array, size_t const index,
+						size_t const value)
 {
 	struct kbase_mem_pool_config *const configs = array;
 
-	if (WARN_ON(!configs) ||
-		WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
+	if (WARN_ON(!configs) || WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
 		return;
 
 	kbase_mem_pool_config_set_max_size(&configs[index], value);
 }
 
-size_t kbase_mem_pool_config_debugfs_max_size(void *const array,
-	size_t const index)
+size_t kbase_mem_pool_config_debugfs_max_size(void *const array, size_t const index)
 {
 	struct kbase_mem_pool_config *const configs = array;
 
-	if (WARN_ON(!configs) ||
-		WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
+	if (WARN_ON(!configs) || WARN_ON(index >= MEMORY_GROUP_MANAGER_NR_GROUPS))
 		return 0;
 
 	return kbase_mem_pool_config_get_max_size(&configs[index]);
@@ -98,25 +89,24 @@ size_t kbase_mem_pool_config_debugfs_max_size(void *const array,
 static int kbase_mem_pool_debugfs_size_show(struct seq_file *sfile, void *data)
 {
 	CSTD_UNUSED(data);
-	return kbase_debugfs_helper_seq_read(sfile,
-		MEMORY_GROUP_MANAGER_NR_GROUPS, kbase_mem_pool_debugfs_size);
+	return kbase_debugfs_helper_seq_read(sfile, MEMORY_GROUP_MANAGER_NR_GROUPS,
+					     kbase_mem_pool_debugfs_size);
 }
 
-static ssize_t kbase_mem_pool_debugfs_write(struct file *file,
-		const char __user *ubuf, size_t count, loff_t *ppos)
+static ssize_t kbase_mem_pool_debugfs_write(struct file *file, const char __user *ubuf,
+					    size_t count, loff_t *ppos)
 {
-	int err;
+	ssize_t err;
 
 	CSTD_UNUSED(ppos);
-	err = kbase_debugfs_helper_seq_write(file, ubuf, count,
-		MEMORY_GROUP_MANAGER_NR_GROUPS, kbase_mem_pool_debugfs_trim);
-	return err ? err : count;
+	err = kbase_debugfs_helper_seq_write(file, ubuf, count, MEMORY_GROUP_MANAGER_NR_GROUPS,
+					     kbase_mem_pool_debugfs_trim);
+	return err ? err : (ssize_t)count;
 }
 
 static int kbase_mem_pool_debugfs_open(struct inode *in, struct file *file)
 {
-	return single_open(file, kbase_mem_pool_debugfs_size_show,
-		in->i_private);
+	return single_open(file, kbase_mem_pool_debugfs_size_show, in->i_private);
 }
 
 static const struct file_operations kbase_mem_pool_debugfs_fops = {
@@ -128,32 +118,27 @@ static const struct file_operations kbase_mem_pool_debugfs_fops = {
 	.release = single_release,
 };
 
-static int kbase_mem_pool_debugfs_max_size_show(struct seq_file *sfile,
-	void *data)
+static int kbase_mem_pool_debugfs_max_size_show(struct seq_file *sfile, void *data)
 {
 	CSTD_UNUSED(data);
-	return kbase_debugfs_helper_seq_read(sfile,
-		MEMORY_GROUP_MANAGER_NR_GROUPS,
-		kbase_mem_pool_debugfs_max_size);
+	return kbase_debugfs_helper_seq_read(sfile, MEMORY_GROUP_MANAGER_NR_GROUPS,
+					     kbase_mem_pool_debugfs_max_size);
 }
 
-static ssize_t kbase_mem_pool_debugfs_max_size_write(struct file *file,
-		const char __user *ubuf, size_t count, loff_t *ppos)
+static ssize_t kbase_mem_pool_debugfs_max_size_write(struct file *file, const char __user *ubuf,
+						     size_t count, loff_t *ppos)
 {
-	int err;
+	ssize_t err;
 
 	CSTD_UNUSED(ppos);
-	err = kbase_debugfs_helper_seq_write(file, ubuf, count,
-		MEMORY_GROUP_MANAGER_NR_GROUPS,
-		kbase_mem_pool_debugfs_set_max_size);
-	return err ? err : count;
+	err = kbase_debugfs_helper_seq_write(file, ubuf, count, MEMORY_GROUP_MANAGER_NR_GROUPS,
+					     kbase_mem_pool_debugfs_set_max_size);
+	return err ? err : (ssize_t)count;
 }
 
-static int kbase_mem_pool_debugfs_max_size_open(struct inode *in,
-	struct file *file)
+static int kbase_mem_pool_debugfs_max_size_open(struct inode *in, struct file *file)
 {
-	return single_open(file, kbase_mem_pool_debugfs_max_size_show,
-		in->i_private);
+	return single_open(file, kbase_mem_pool_debugfs_max_size_show, in->i_private);
 }
 
 static const struct file_operations kbase_mem_pool_debugfs_max_size_fops = {
@@ -165,20 +150,19 @@ static const struct file_operations kbase_mem_pool_debugfs_max_size_fops = {
 	.release = single_release,
 };
 
-void kbase_mem_pool_debugfs_init(struct dentry *parent,
-		struct kbase_context *kctx)
+void kbase_mem_pool_debugfs_init(struct dentry *parent, struct kbase_context *kctx)
 {
 	const mode_t mode = 0644;
 
-	debugfs_create_file("mem_pool_size", mode, parent,
-		&kctx->mem_pools.small, &kbase_mem_pool_debugfs_fops);
+	debugfs_create_file("mem_pool_size", mode, parent, &kctx->mem_pools.small,
+			    &kbase_mem_pool_debugfs_fops);
 
-	debugfs_create_file("mem_pool_max_size", mode, parent,
-		&kctx->mem_pools.small, &kbase_mem_pool_debugfs_max_size_fops);
+	debugfs_create_file("mem_pool_max_size", mode, parent, &kctx->mem_pools.small,
+			    &kbase_mem_pool_debugfs_max_size_fops);
 
-	debugfs_create_file("lp_mem_pool_size", mode, parent,
-		&kctx->mem_pools.large, &kbase_mem_pool_debugfs_fops);
+	debugfs_create_file("lp_mem_pool_size", mode, parent, &kctx->mem_pools.large,
+			    &kbase_mem_pool_debugfs_fops);
 
-	debugfs_create_file("lp_mem_pool_max_size", mode, parent,
-		&kctx->mem_pools.large, &kbase_mem_pool_debugfs_max_size_fops);
+	debugfs_create_file("lp_mem_pool_max_size", mode, parent, &kctx->mem_pools.large,
+			    &kbase_mem_pool_debugfs_max_size_fops);
 }
